+++ packages/sdk4337/src/types.ts
@@ -1,6 +1,7 @@
 import { ERC20OrNativeTokenMetadata } from '@0xsodium/utils';
 import { BigNumber } from 'ethers';
 export type PaymasterInfo = {
+    id: string
     token: ERC20OrNativeTokenMetadata
 
     // wei