+++ packages/bundler/CHANGELOG.md
@@ -1,5 +1,71 @@
 # @0xsodium/bundler
 
+## 0.9.0
+
+### Minor Changes
+
+- fix iframe provider bug
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/utils@0.9.0
+
+## 0.8.0
+
+### Minor Changes
+
+- Support iframe provider
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/utils@0.8.0
+
+## 0.7.0
+
+### Minor Changes
+
+- Support paymaster
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/utils@0.7.0
+
+## 0.6.0
+
+### Minor Changes
+
+- fix bugs
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/utils@0.6.0
+
+## 0.5.0
+
+### Minor Changes
+
+- fix bugs
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/utils@0.5.0
+
+## 0.4.0
+
+### Minor Changes
+
+- fix bugs
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/utils@0.4.0
+
 ## 0.3.0
 
 ### Minor Changes